make migration check work with root dir start scripts
